dvips pTeX extensions, based on patches sent by Akira
git-svn-id: svn://tug.org/texlive/trunk/Build/source@18107 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipsk/repack.c
@@ -422,7 +422,7 @@ repack(register chardesctype *cp)
       mraster[--i] = 0;
    i = startbytes + unpack(p, (halfword *)mraster, (halfword)width,
                 (halfword)height,  *(cp->packptr));
-   dochar(mraster, (width + 7) >> 3, height);
+   dochar(mraster, (short)((width + 7) >> 3), (short)height);
    if (smallchar) {
       addts((unsigned char)0);
       addts((unsigned char)0);
